add oled and encoder function to configutator
